Added tons of documentation and debuggoing logging to intervalListIntersections

+++ LargetrParenthesis/DotProductTwoSparseVector.h
@@ -53,7 +53,7 @@ namespace DotProductTwoSparseVector
 	public:
 		SparseVector(const vector<int>& nums)
 		{
-			for (int i = 0; i < nums.size(); i++)
+			for (size_t i = 0; i < nums.size(); i++)
 			{
 				if (nums[i] != 0)
 				{
